Move organization update notifications to update.js
