+++ backend/services/partnerNetwork/update.js
@@ -1,5 +1,9 @@
-const { GDBServiceProviderModel } = require("../../models");
-const { GDBOrganizationModel } = require("../../models/organization");
+const {formatLocation} = require("../../helpers/location");
+const {sanitize} = require("../../helpers/sanitizer");
+const {GDBServiceProviderModel, GDBAddressModel} = require("../../models");
+const {GDBOrganizationModel} = require("../../models/organization");
+const {getIndividualsInClass} = require("../dynamicForm");
+const {createNotificationHelper} = require("../notification/notification");
 
 async function sendPartnerUpdateNotification(req, partnerId) {
   const { fetchSingleGenericHelper } = require("../genericData");
@@ -51,6 +55,19 @@ async function sendPartnerUpdateNotification(req, partnerId) {
   }
 }
 
+async function printAddress(address) {
+  const streetTypes = await getIndividualsInClass('ic:StreetType');
+  const streetDirections = await getIndividualsInClass('ic:StreetDirection');
+  const states = await getIndividualsInClass('schema:State');
+  const addressInfo = {streetTypes, streetDirections, states};
+
+  if (typeof address === 'string') {
+    address = await GDBAddressModel.findOne({_id: address.split('_')[1]});
+  }
+
+  return sanitize(formatLocation(address, addressInfo));
+}
+
 async function receivePartnerUpdateNotification(req, res, next) {
   const { fetchOrganizationHelper, deleteOrganizationHelper, updateOrganizationHelper } = require(".");
 
@@ -76,10 +93,30 @@ async function receivePartnerUpdateNotification(req, res, next) {
       const partnerServiceProvider = await GDBServiceProviderModel
         .findOne({organization: {_id: partnerOrganization._id}});
       if (!partnerData.organization || Object.keys(partnerData.organization).length === 0) {
-        await deleteOrganizationHelper(partnerServiceProvider._id, partnerData);
+        const {partnerJson} = await deleteOrganizationHelper(partnerServiceProvider._id, partnerData);
+        // Notify the user of the deleted partner organization
+        createNotificationHelper({
+          name: 'A partner deployment deleted its home organization',
+          description: `<div><p>${sanitize(partnerJson.name)}, one of your partners, just deleted its home organization.</p>`
+            + `<p>To reconnect with them, ask them to set up a home organization again and then create a partner `
+            + `organization for that partner.</p>`
+            + `<p>For reference, here are the details of the partner:</p>`
+            + `<dl>`
+            + (partnerJson.description?.length > 0 ? `<dt>Description:</dt><dd>${sanitize(partnerJson.description)}</dd>` : '')
+            + (partnerJson.address ? `<dt>Address:</dt><dd>${await printAddress(partnerJson.address)}</dd>` : '')
+            + `<dt>Endpoint URL:</dt><dd>${partnerJson.endpointUrl}</dd>`
+            + `<dt>Endpoint Port:</dt><dd>${partnerJson.endpointPort}</dd>`
+            + `</dl></div>`
+        });
         return res.status(200).json({success: true});
       } else {
-        await updateOrganizationHelper(partnerServiceProvider._id, partnerData);
+        const {partnerJson} = await updateOrganizationHelper(partnerServiceProvider._id, partnerData);
+        // Notify the user of the updated partner organization
+        createNotificationHelper({
+          name: 'A partner deployment updated its home organization',
+          description: `<a href="/providers/organization/${partnerJson._id}">${sanitize(partnerJson.name)}</a>, one of your `
+            + `partners, just updated its home organization.`
+        });
         return res.status(200).json({success: true});
       }
     } else {
